fix: Adapt parameters. First flying glider!
With some tuning of the hydrodynamic parameters and the mass
this glider actually flies up and down when the bladder is deflaed /
inflated!
Note: the parameters are not yeat meant to be an accurate re-
presentation of Nautilus, but they are roughly in the right ballpark
and now tuned with respect to each other so that the behaviour is as
expected. When starting the simulation, the bladder starts off empty.
ACU not yet implemented.